feat: Very rough initial prototype
Implements a rough outline of how the backup utility will be built. The
tool consists of a few key components and dependencies.

Clap is used to build the command line interface including arguments,
default values, validation of arguments, etc. The library is one of the
more used and support most what is needed to build the CLI.

Jmap_client is used to interface with the JMAP servers. It handles
setting up queries and other requests as well as error handling. It
makes it trivial to fetch mailboxes, emails, blobs, sieves, etc.

Apache OpenDAL is used to abstract away the storage to use for backup.
On command startup we initialize the requested (from command line)
OpenDAL operator, be it Fs (filesystem), S3, or other. This operator is
then passed to the backup function which can use the operator to write
the backup files.

Keyring_rs is used to store sensitive values including passwords and
tokens. Not sure how to handle the OpenDAL secrets as there are many
ways to provide these depending on the backend. Might need to think of
something more clever here. The user should run the command
interactively at least once to set the passwords/tokens, and then the
command can be run without being prompted again. This will be useful
when running in cron, etc. Storing credentials in keyring is better than
having them in plaintext config files.

The program makes use of async support in Rust, but most of the code is
currently structured in a way where all logic executes synchronously.
This means we don't get the benefit of async for performance gains.

JMAP allows batching reads, so currently writing will most likely be the
bottleneck. Need to study async Rust in more detail and find a way to
improve the efficiency of the code.

The code is currently not incremental, it just fetches every mailbox and
email from the server (- body which is separate blobs in JMAP). The
logic for fetching BLOBs should be quite similar though.