feat: Config file support and refactoring
Implements support for configuration via files and environment
variables courtesy config crate. Users can opt to specify a specific
file to use as an option to the postkasse command or allow it to
fallback to the os specific location to look for a postkasse.toml file.

Secrets are now handled by first looking in the config file, then next
in the keyring. If no secret is found it prompts the user for it.

Refactors the backup logic to separate logic relating to mailboxes and
emails in separate files. Before starting the backup it reads the
progress files to figure out where it last left off. It then continues
from that point onward.