feat: Initial search indexing support
Adds initial search indexing support and a search command to Postkasse.
Search is, as teased in the README, done using Tantivy. The code adds
search index fields for things like the jmap email id, blob id, subject,
from, to, and cc names and emails as well as the text body.

The shorter values are stored to be used in search result lists. The
longer values like the text body is not stored as Postkasse can just
fetch the email on demand.

In order to get the text body we use the mail_parser create by Stalwart
Labs which allows us to parse the blob downloaded via JMAP server into a
Message value where we can access the text body. This way Postkasse does
not need to spam the JMAP server to get the text body of an email, thus
reducing load on the JMAP server.

Rayon is used to parse and index emails in parallell thus allowing
faster backup of emails. It runs once emails and blobs have been written
to the OpenDAL operator.

The current backup implementation has lots of room for performance
improvements, but the approach should be performant enough for something
that will be a rather small inremental job.