refactor: Make Progressable trait for progress tracking
Instead of relying directly on the ProgressBar type in the core the
backup functions should instead rely on a trait. Then for the CLI use
case where a progress bar is used the code implements the Progressable
trait for the ProgressBar struct so we can use it to render both the
progress bar and track progress.